+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `SqliteDictFormat` for datasets.
+- Added `pre_epoch()` and `post_epoch()` callback methods to PyTorch `TrainCallback`.
 
 ### Changed